package everything into 3 containers
