+++ dockerfiles/dcpp_scrape.Dockerfile
@@ -60,34 +60,6 @@ RUN sudo apt-get -y install \
               python3-yaml
 RUN pipx install wllvm
 
-#####################
-# Installing PhASAR #
-#####################
-
-WORKDIR /usr/src/
-RUN git clone https://github.com/secure-software-engineering/phasar.git
-WORKDIR /usr/src/phasar
-
-ENV CC=/usr/bin/clang-14
-ENV CXX=/usr/bin/clang++-14
-
-RUN git submodule init
-RUN git submodule update
-RUN mkdir -p build && cd build && \
-          cmake .. \
-            -DCMAKE_BUILD_TYPE=RelWithDebInfo \
-            -DPHASAR_TARGET_ARCH="" \
-            -DCMAKE_CXX_COMPILER=$CXX \
-            -DPHASAR_BUILD_UNITTESTS=OFF \
-            -DPHASAR_BUILD_IR=OFF \
-            -DPHASAR_BUILD_DOC=ON \
-            -G Ninja && \
-          ninja -j $(nproc) && \
-          sudo ninja install
-
-# putting PhASAR in PATH
-ENV PATH="$PATH:/usr/local/lib:/usr/local/include/"
-
 # installing go (for gllvm)
 RUN apt-get -y update && \
     apt-get -y install golang-go
@@ -125,20 +97,14 @@ RUN git clone https://github.com/conan-io/conan-center-index
 RUN conan remote add conan-index /conan-center-index
 RUN conan profile detect
 
-# logging library for c++
-RUN sudo apt install -y libspdlog-dev
+####################
+# Getting DelpiCpp #
+####################
 
+WORKDIR /
+COPY ../scraper/ /scraper/
 
-#####################
-# Building DelpiCpp #
-#####################
+RUN mkdir /volumes/
+RUN touch /volumes/log.log
 
-WORKDIR /workspaces/
-
-RUN git clone https://github.com/sse-labs/delphiCpp.git
-RUN mv /workspaces/delphiCpp/ /workspaces/DelphiCpp/
-WORKDIR /workspaces/DelphiCpp/build/
-RUN cmake -DCMAKE_BUILD_TYPE=Release /workspaces/DelphiCpp/CMakeLists.txt
-RUN make
-
-ENTRYPOINT [ "python3", "/workspaces/DelphiCpp/partial_extract.py", "-c", "/workspaces/DelphiCpp/example_configs/config.json", "-pi", "/workspaces/DelphiCpp/bitcode/packages.json", "-fm", "/my_vol/res_fm.json", "-s", "/my_vol/eval_stats" ]
+ENTRYPOINT [ "python3", "-m", "scraper", "-o", "/volumes/packages/", "-l", "/volumes/log.log" ]
