+++ entrypoints/vol_paths.py
@@ -1,0 +1,9 @@
+paths = {
+  'pkgs': '/volumes/packages/',
+  'conf': '/volumes/config.json',
+  'filters': '/volumes/filters.json',
+  'ftr_map': '/volumes/ftr_map.json',
+  'filtered_map': '/volumes/fltrd_map.json',
+  'log_file': '/volumes/log.log',
+  'log_dir': '/volumes/logs/',
+}
